Split test_error_messages.py along source-module lines
The suite names test modules after the source module they test, so the
from_file test moves to test_data_models.py and the
from_pre_registered_model test moves to test_mapping_tables.py.
